refactor(api): migrate database layer from PostgreSQL/pg-core to Cloudflare D1/sqlite-core
33 tables converted to sqlite-core; raw-SQL PG constructs replaced with SQLite equivalents; better-auth adapter provider changed from pg to sqlite; drizzle-orm bumped to 0.45.2; db/index.ts now exports createDb(d1) and initD1, with a throwing placeholder pending the Cloudflare entrypoint.

+++ apps/api/src/domains/assessments/services/AssessmentsService.ts
@@ -232,7 +232,7 @@ export class AssessmentsService {
     const { page, limit, offset } = getPagination(query, 10, 100);
 
     const cycleCountResult = await db
-      .select({ count: sql<number>`count(distinct ${AssessmentTable.cycle_start_date})::int` })
+      .select({ count: sql<number>`count(distinct ${AssessmentTable.cycle_start_date})` })
       .from(AssessmentTable)
       .where(eq(AssessmentTable.student_id, studentId));
